requirements.txt
-i https://pypi.org/simple
pyopengl==3.1.5
pyqt5-sip==12.8.0
pyqt5==5.15.0
